+++ config/environments/production.rb
@@ -23,7 +23,7 @@ Rails.application.configure do
   config.public_file_server.enabled = ENV['RAILS_SERVE_STATIC_FILES'].present?
 
   # Compress JavaScripts and CSS.
-  config.assets.js_compressor = :uglifier
+  config.assets.js_compressor = Uglifier.new(harmony: true)
   # config.assets.css_compressor = :sass
 
   # Do not fallback to assets pipeline if a precompiled asset is missed.
@@ -104,7 +104,7 @@ Rails.application.configure do
 
   config.action_dispatch.default_headers.merge!(
     # Value given by https://www.gov.uk/service-manual/technology/using-https
-    'Strict-Transport-Security' => 'max-age=31536000, includeSubDomains; preload;',
+    'Strict-Transport-Security' => 'max-age=31536000; includeSubDomains; preload',
 
     'Expect-CT' => 'enforce, max-age=10'
   )