+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## [release-77] - 2022-01-27
+
+- RMI-378: Cookie banner
+- RMI-379: Cookie setting and policy pages
+- RMI-478: Update to HTST config
+
 ## [release-76] - 2021-11-25
 
 - RMI-418: Guard against incorrectly issuing credit notes.
@@ -406,6 +412,7 @@
 
 Initial release
 
+[release-77]: https://github.com/Crown-Commercial-Service/DataSubmissionService/compare/release-76...release-77
 [release-76]: https://github.com/Crown-Commercial-Service/DataSubmissionService/compare/release-75...release-76
 [release-75]: https://github.com/Crown-Commercial-Service/DataSubmissionService/compare/release-74...release-75
 [release-74]: https://github.com/Crown-Commercial-Service/DataSubmissionService/compare/release-73...release-74